+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- `omc4py.v_1_24` Support of _OpenModelica 1.24.*_
+
 ## [0.3.2] - 2024-01-31
 
 ### Fixed